requirements.txt
fastapi[all]
asyncpg
python-dotenv
alembic
psycopg2
python-socketio
bcrypt
